+++ Controller/Router.php
@@ -4,12 +4,20 @@ namespace Conekta\Payments\Controller;
 
 use Conekta\Payments\Helper\{Data, Data as ConektaHelper};
 use Conekta\Payments\Logger\Logger as ConektaLogger;
-use Magento\Framework\App\{ActionFactory, ActionInterface, RequestInterface, ResponseInterface, RouterInterface};
+use Magento\Framework\App\ActionFactory;
+use Magento\Framework\App\RequestInterface;
+use Magento\Framework\App\ResponseInterface;
+use Magento\Framework\App\RouterInterface;
 use Magento\Framework\Url;
 
+/**
+ * Class Router
+ */
 class Router implements RouterInterface
 {
     /**
+     * Router construct
+     *
      * @param ActionFactory $actionFactory
      * @param ResponseInterface $_response
      * @param Data $conektaHelper
@@ -28,6 +36,7 @@ class Router implements RouterInterface
      *
      * @param RequestInterface $request
      * @return void
+     * @throws \Magento\Framework\Exception\NoSuchEntityException
      */
     public function match(RequestInterface $request): void
     {